nfc_backend/requirements.txt
django
daphne
channels
djangorestframework
django-cors-headers
pyscard
asgiref
